% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edld652-package.R
\docType{package}
\name{edld652-package}
\alias{edld652}
\alias{edld652-package}
\title{edld652: Course Package for EDLD 652: Data Visualization for Educational Data Science}
\description{
This is a package designed for retrieving data for the EDLD 652 course taught at the University of Oregon.
}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu} (\href{https://orcid.org/0000-0003-4699-4680}{ORCID})

}
\keyword{internal}
