% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Function to retrieve course data}
\usage{
get_data(name)
}
\arguments{
\item{name}{The name of the file you want to import. Should be a string
returned by \code{\link[=list_datasets]{list_datasets()}}.}
}
\description{
Note there are many possible data files to pull. To see all possible data
files, please run \code{list_datasets()}. These files link together via keys. You
should be able to join multiple files to get a more complete dataset
from which to visualize your results.
}
