% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_documentation.R
\name{get_documentation}
\alias{get_documentation}
\title{Downloads a Microsoft Word file containing the data documentation for a given dataset}
\usage{
get_documentation(name)
}
\arguments{
\item{name}{The name of the file you want to import. Should be a string
returned by \code{\link[=list_datasets]{list_datasets()}}.}
}
\description{
Downloads a Microsoft Word file containing the data documentation for a given dataset
}
