% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_fiscal_data}
\alias{get_fiscal_data}
\title{Imports all fiscal data available}
\usage{
get_fiscal_data()
}
\description{
Returns a list of all fiscal data.
}
\details{
Note -  this may not work if you don't have a lot of memory, and it will
take a considerable amount of time regardless. Be prepared to wait several
minutes.
}
\examples{
\dontrun{
l <- get_fiscal_data()

# fiscal dataframe from 2010
l$sy2010

l$sy2018

}
}
