% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{get_key}
\alias{get_key}
\title{Access your key from .Renviron}
\usage{
get_key()
}
\description{
This function is primarily used as the default argument for
\code{\link[=get_data]{get_data()}}, but can be run at any time to access your
specific key. Note that you should first set your key with
\code{\link[=set_key]{set_key()}}.
}
\examples{
\dontrun{
library(edld652)
get_key()
}

}
