% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{list_datasets}
\alias{list_datasets}
\title{List  the available dataset}
\usage{
list_datasets()
}
\description{
This function will list all possible datasets that are available for you to
work with in this class. You can import any of these into R using by
supplying the name of the dataset (as a string) to the \code{\link[=get_data]{get_data()}}
function.
}
